Make I/O error msgs more helpful
I/O errors originate from Python (because that's where our streams come
from), so it's a good idea to include the Python tracebacks in them.

The format is quite ugly ATM (basically repr() instead of str() like
output) because it goes through Rust's IO error type, but I guess fixing
that would be more trouble than it's worth for now.